requirements.txt
Django
django-ckeditor
django-cleanup
django-filter
django-jazzmin
django-phonenumber-field
django-cors-headers
djangorestframework
django-rest-registration
drf-yasg
django-resized
phonenumbers
Pillow
python-decouple
